starting to work on favourites page
